Messages appear on Thread
Made changes to the model and view. Threads are queried using user objects and not usernames. Just as before, if two thread objects exists with the same 2 users but different first and second position then the one created first is sent.
Messages added in from the Admin page appears in the thread. Will see now about sending messages to the consumer and saving it to the backend.