Quick URL Redirect from Index Page
Made a change previously to accomodate for the messaging websocket. Having the feed socket url as '' caused the messaging socket to redirect connections to the PostConsumer as opposed to the ChatConsumer.
This fix allows us to redirect users from the website to the feed page and keeps socket connections safe.